+++ src/BGS/ch/yumblie.c
@@ -115,7 +115,7 @@ Actor *chYumblie_draw(ActorMarker *this, Gfx **gfx, Mtx** mtx, Vtx **vtx){
     }
 
     modelRender_setPreDrawCallback((GenFunction_1) actor_predrawMethod, (intptr_t)thisActor);
-    modelRender_setPostDrawCallback((GenFunction_1) actor_postdrawMethod, (s32)this);
+    modelRender_setPostDrawCallback((GenFunction_1) actor_postdrawMethod, (intptr_t)this);
     actor_position[0] = thisActor->position_x;
     actor_position[1] = thisActor->position_y + local->animTimer * 75.0f;
     actor_position[2] = thisActor->position_z;
