+++ src/GV/ch/gobirock.c
@@ -69,7 +69,7 @@ Actor *chGobiRock_draw(ActorMarker *this_marker, Gfx **gfx, Mtx **mtx, Vtx **vtx
     sp3C[0] = this->pitch;
     sp3C[1] = this->yaw;
     sp3C[2] = this->roll;
-    modelRender_setPreDrawCallback((GenFunction_1)actor_predrawMethod, (s32) this);
+    modelRender_setPreDrawCallback((GenFunction_1)actor_predrawMethod, (intptr_t)this);
     modelRender_setRefPoints(func_80329934());
     modelRender_draw(gfx, mtx, this->position, sp3C, 1.0f, NULL,marker_loadModelBin(this_marker));
     vec3fArray_get_vec3f(func_80329934(), 5, local->unk0);
