+++ src/GV/ch/histup.c
@@ -118,7 +118,7 @@ void func_8038DBDC(Actor *this){
                 func_8038DB4C(0);
                 this->unk1C[0] = 0.0f;
             }//L8038DC90
-            if(func_8038E178() != (s32)this->unk1C[0] || volatileFlag_get(VOLATILE_FLAG_C1_IN_FINAL_CHARACTER_PARADE)){
+            if(func_8038E178() != (intptr_t)this->unk1C[0] || volatileFlag_get(VOLATILE_FLAG_C1_IN_FINAL_CHARACTER_PARADE)){
                 if(15.0f <= this->lifetime_value){
                     subaddie_set_state_with_direction(this, 2, 0.01f, 1);
                     anctrl_setPlaybackType(this->anctrl, ANIMCTRL_ONCE);
