+++ src/core2/ch/badShad.c
@@ -4,7 +4,7 @@
 #include "variables.h"
 
 
-extern void __spawnQueue_add_2(void (*)(s32, s32), s32, s32);
+extern void __spawnQueue_add_2(void (*)(intptr_t, intptr_t), s32, s32);
 
 Actor *chBadShad_draw(ActorMarker *this, Gfx **gdl, Mtx **mptr, Vtx **arg3);
 void chBadShad_update(Actor *this);
@@ -139,7 +139,7 @@ void func_802D729C(Actor *actor, f32 arg1){
         return;
     }
     if(!actor->unk104){
-        __spawnQueue_add_2((GenFunction_2)func_802D71A0, (s32) actor->marker, reinterpret_cast(intptr_t, arg1));
+        __spawnQueue_add_2((GenFunction_2)func_802D71A0, (intptr_t)actor->marker, reinterpret_cast(intptr_t, arg1));
         return;
     }
