+++ src/core2/ch/gravestone.c
@@ -59,9 +59,9 @@ Actor *chgravestone_draw(ActorMarker *marker, Gfx **gfx, Mtx **mtx, Vtx **vtx) {
 
     this = marker_getActorAndRotation(marker, rotation);
     local = (ActorLocal_core2_D7D10 *)&this->local;
-    modelRender_setAppendageVisibility(3, (s32) this->velocity[1]);
+    modelRender_setAppendageVisibility(3, (intptr_t)this->velocity[1]);
     modelRender_setAppendageVisibility(4, local->unk0);
-    modelRender_setPreDrawCallback( (GenFunction_1)actor_predrawMethod, (s32)this);
+    modelRender_setPreDrawCallback( (GenFunction_1)actor_predrawMethod, (intptr_t)this);
     modelRender_setPostDrawCallback((GenFunction_1)actor_postdrawMethod, (s32)marker);
     modelRender_draw(gfx, mtx, this->position, rotation, this->scale, D_80373008, func_803257B4(marker));
     return this;
